backend/requirements.txt
# --- API y Servidor ---
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1

# --- OCR y Procesamiento de Imágenes ---
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.3.0

# --- Procesamiento de Texto y Machine Learning ---
spacy==3.7.4
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# --- Base de Datos (SQLite/SQLAlchemy) ---
SQLAlchemy==2.0.29

# --- Testing ---
pytest==8.1.1

# --- Utilidades PDF ---
PyMuPDF==1.24.1
